// frontend/src/types/generated/models/BaselineComparison.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DiversityTrend } from "./DiversityTrend";

export type BaselineComparison = { overallDiff: number, lexicalDiff: number, structuralDiff: number, semanticDiff: number, trend: DiversityTrend, };